# attached_assets-4/attached_assets/2/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clientum CRM/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /openrouter/conversations:
    get:
      operationId: listOpenrouterConversations
      tags: [openrouter]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterConversation"
    post:
      operationId: createOpenrouterConversation
      tags: [openrouter]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenrouterConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversation"
  /openrouter/conversations/{id}:
    get:
      operationId: getOpenrouterConversation
      tags: [openrouter]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversationWithMessages"
        "404":
          description: Not found
    delete:
      operationId: deleteOpenrouterConversation
      tags: [openrouter]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /openrouter/conversations/{id}/messages:
    get:
      operationId: listOpenrouterMessages
      tags: [openrouter]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterMessage"
    post:
      operationId: sendOpenrouterMessage
      tags: [openrouter]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenrouterMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, title, stage, value, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        stage:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadInput:
      type: object
      required: [title, stage, value]
      properties:
        title:
          type: string
        contactId:
          type: integer
        companyId:
          type: integer
        stage:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        source:
          type: string
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        title:
          type: string
        contactId:
          type: integer
        companyId:
          type: integer
        stage:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        source:
          type: string
        notes:
          type: string
...
    OpenrouterConversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
    OpenrouterMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    OpenrouterConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    OpenrouterMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    OpenrouterConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenrouterMessage"
...