# attached_assets-4/attached_assets/2/lib/api-zod/src/generated/types/openrouterMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clientum CRM/ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenrouterMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: Date;
}