# attached_assets-4/attached_assets/Chat-Export-1s/lib/api-zod/src/generated/types/openrouterConversation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clientum CRM/ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenrouterConversation {
  id: number;
  title: string;
  createdAt: Date;
}