# attached_assets-4/attached_assets/Chat-Export-1s/lib/api-zod/src/generated/types/openrouterConversationWithMessages.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clientum CRM/ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { OpenrouterMessage } from './openrouterMessage';

export interface OpenrouterConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  messages: OpenrouterMessage[];
}